+++ test/selenium/test_initial_setup.py
@@ -285,6 +285,9 @@ class TestInitialSetup(SeleniumTest):
         # Check that we are re-directed to setup page
         self.assert_equals(lambda: self.selenium_instance.current_url, self.get_url('/initial-setup'))
 
+        # Check page title
+        assert self.selenium_instance.title == 'Initial Setup - Terrareg'
+
         # STEP 1 - Auth environment variables
         self._test_auth_vars_step()