+++ test/selenium/test_module_provider.py
@@ -20,6 +20,7 @@ class TestModuleProvider(SeleniumTest):
         cls._api_version_publish_mock = mock.Mock(return_value={'status': 'Success'})
         cls._config_publish_api_keys_mock = mock.patch('terrareg.config.Config.PUBLISH_API_KEYS', [])
         cls._config_allow_custom_repo_urls_module_provider = mock.patch('terrareg.config.Config.ALLOW_CUSTOM_GIT_URL_MODULE_PROVIDER', True)
+        cls._config_allow_custom_repo_urls_module_version = mock.patch('terrareg.config.Config.ALLOW_CUSTOM_GIT_URL_MODULE_VERSION', True)
 
         cls.register_patch(mock.patch('terrareg.config.Config.ADMIN_AUTHENTICATION_TOKEN', 'unittest-password'))
         cls.register_patch(mock.patch('terrareg.config.Config.SECRET_KEY', '354867a669ef58d17d0513a0f3d02f4403354915139422a8931661a3dbccdffe'))
@@ -27,8 +28,20 @@ class TestModuleProvider(SeleniumTest):
         cls.register_patch(mock.patch('terrareg.server.ApiTerraregModuleVersionPublish._post', cls._api_version_publish_mock))
         cls.register_patch(cls._config_publish_api_keys_mock)
         cls.register_patch(cls._config_allow_custom_repo_urls_module_provider)
+        cls.register_patch(cls._config_allow_custom_repo_urls_module_version)
 
         super(TestModuleProvider, cls).setup_class()
+
+    @pytest.mark.parametrize('url,expected_title', [
+        ('/modules/moduledetails/noversion/testprovider', 'moduledetails/noversion/testprovider - Terrareg'),
+        ('/modules/moduledetails/fullypopulated/testprovider/1.5.0', 'moduledetails/fullypopulated/testprovider - Terrareg'),
+        ('/modules/moduledetails/fullypopulated/testprovider/1.5.0/example/examples/test-example', 'moduledetails/fullypopulated/testprovider/examples/test-example - Terrareg'),
+        ('/modules/moduledetails/fullypopulated/testprovider/1.5.0/submodule/modules/example-submodule1', 'moduledetails/fullypopulated/testprovider/modules/example-submodule1 - Terrareg'),
+    ])
+    def test_page_titles(self, url, expected_title):
+        """Check page titles on pages."""
+        self.selenium_instance.get(self.get_url(url))
+        self.assert_equals(lambda: self.selenium_instance.title, expected_title)
 
     def _get_settings_field_by_label(self, label):
         """Return input element by label."""
@@ -96,8 +109,13 @@ class TestModuleProvider(SeleniumTest):
     ])
     def test_module_with_versions(self, attribute_to_remove, related_element, expect_displayed, expected_display_value):
         """Test page functionality on a module with versions."""
+        # Remove custom browse/base URL from module provider
+        module_provider = ModuleProvider(Module(Namespace('moduledetails'), 'fullypopulated'), 'testprovider')
+        original_provider_repo_base_url_template = module_provider._get_db_row()['repo_base_url_template']
+        original_provider_repo_browse_url_template = module_provider._get_db_row()['repo_browse_url_template']
+        module_provider.update_attributes(repo_base_url_template=None, repo_browse_url_template=None)
+
         if attribute_to_remove:
-            module_provider = ModuleProvider(Module(Namespace('moduledetails'), 'fullypopulated'), 'testprovider')
             module_version = ModuleVersion.get(module_provider, '1.5.0')
             original_value = module_version._get_db_row()[attribute_to_remove]
             module_version.update_attributes(**{attribute_to_remove: None})
@@ -166,6 +184,9 @@ module "fullypopulated" {{
             self.wait_for_element(By.ID, 'no-version-available', ensure_displayed=False).is_displayed == False
 
         finally:
+            module_provider.update_attributes(
+                repo_base_url_template=original_provider_repo_base_url_template,
+                repo_browse_url_template=original_provider_repo_browse_url_template)
             if attribute_to_remove:
                 module_version.update_attributes(**{attribute_to_remove: original_value})
 
@@ -217,6 +238,446 @@ module "fullypopulated" {{
         security_issues = self.wait_for_element(By.ID, 'security-issues')
         assert security_issues.is_displayed() == True
         assert security_issues.text == '1 Security issues'
+
+
+    @pytest.mark.parametrize('url,git_provider_id,module_provider_browse_url_template,module_provider_base_url_template,module_version_browse_url_template,module_version_base_url_template,allow_custom_git_urls_module_provider,allow_custom_git_urls_module_version,expected_source', [
+        # Test with all URLs configured and all custom URLs allowed
+        # - base URL
+        (
+            '/modules/moduledetails/fullypopulated/testprovider',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://module-version.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/suffix'
+        ),
+        # - non-latest version, defaults to using module provider, as the module version
+        # has been configured in the latest version
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.2.0',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://module-provider.com/moduledetails/fullypopulated-testprovider/browse/1.2.0/suffix'
+        ),
+        # - example
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/example/examples/test-example',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://module-version.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/examples/test-examplesuffix'
+        ),
+        # - submodule
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/submodule/modules/example-submodule1',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://module-version.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/modules/example-submodule1suffix'
+        ),
+        # Test with all URLs configured and custom module version URLs disabled
+        # - base URL
+        (
+            '/modules/moduledetails/fullypopulated/testprovider',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            False,
+            'https://module-provider.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/suffix'
+        ),
+        # - example
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/example/examples/test-example',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            False,
+            'https://module-provider.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/examples/test-examplesuffix'
+        ),
+        # - submodule
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/submodule/modules/example-submodule1',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            False,
+            'https://module-provider.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/modules/example-submodule1suffix'
+        ),
+        # Test with all URLs configured and all custom module URLs disabled
+        # - base URL
+        (
+            '/modules/moduledetails/fullypopulated/testprovider',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            False,
+            False,
+            'https://localhost.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/'
+        ),
+        # - example
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/example/examples/test-example',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            False,
+            False,
+            'https://localhost.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/examples/test-example'
+        ),
+        # - submodule
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/submodule/modules/example-submodule1',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            False,
+            False,
+            'https://localhost.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/modules/example-submodule1'
+        ),
+        # Test with all URLs configured except module version browse and all custom URLs enabled
+        # - base URL
+        (
+            '/modules/moduledetails/fullypopulated/testprovider',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://module-provider.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/suffix'
+        ),
+        # - example
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/example/examples/test-example',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://module-provider.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/examples/test-examplesuffix'
+        ),
+        # - submodule
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/submodule/modules/example-submodule1',
+            1,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse/{tag}/{path}suffix',
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://module-provider.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/modules/example-submodule1suffix'
+        ),
+        # Test with all URLs configured except module version/provider browse and all custom URLs enabled
+        # - base URL
+        (
+            '/modules/moduledetails/fullypopulated/testprovider',
+            1,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://localhost.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/'
+        ),
+        # - example
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/example/examples/test-example',
+            1,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://localhost.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/examples/test-example'
+        ),
+        # - submodule
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/submodule/modules/example-submodule1',
+            1,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://localhost.com/moduledetails/fullypopulated-testprovider/browse/1.5.0/modules/example-submodule1'
+        ),
+        # Test with no browse URLs configured and all custom URLs enabled, testing revert to base URL
+        # - base URL
+        (
+            '/modules/moduledetails/fullypopulated/testprovider',
+            4,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://module-version.com/moduledetails/fullypopulated-testprovider/browse'
+        ),
+        # - example
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/example/examples/test-example',
+            4,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://module-version.com/moduledetails/fullypopulated-testprovider/browse'
+        ),
+        # - submodule
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/submodule/modules/example-submodule1',
+            4,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            True,
+            'https://module-version.com/moduledetails/fullypopulated-testprovider/browse'
+        ),
+        # Test with no browse URLs configured and module version custom URLs disabled, testing revert to base URL
+        # - base URL
+        (
+            '/modules/moduledetails/fullypopulated/testprovider',
+            4,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            False,
+            'https://module-provider.com/moduledetails/fullypopulated-testprovider/browse'
+        ),
+        # - example
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/example/examples/test-example',
+            4,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            False,
+            'https://module-provider.com/moduledetails/fullypopulated-testprovider/browse'
+        ),
+        # - submodule
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/submodule/modules/example-submodule1',
+            4,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            True,
+            False,
+            'https://module-provider.com/moduledetails/fullypopulated-testprovider/browse'
+        ),
+        # Test with no browse URLs configured and all custom URLs disabled, testing revert to base URL
+        # - base URL
+        (
+            '/modules/moduledetails/fullypopulated/testprovider',
+            4,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            False,
+            False,
+            'https://base-url.com/moduledetails/fullypopulated-testprovider'
+        ),
+        # - example
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/example/examples/test-example',
+            4,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            False,
+            False,
+            'https://base-url.com/moduledetails/fullypopulated-testprovider'
+        ),
+        # - submodule
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/submodule/modules/example-submodule1',
+            4,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            False,
+            False,
+            'https://base-url.com/moduledetails/fullypopulated-testprovider'
+        ),
+        # Test with no browse URLs configured and all custom URLs disabled and not git provider
+        # - base URL
+        (
+            '/modules/moduledetails/fullypopulated/testprovider',
+            None,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            False,
+            False,
+            None
+        ),
+        # - example
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/example/examples/test-example',
+            None,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            False,
+            False,
+            None
+        ),
+        # - submodule
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/submodule/modules/example-submodule1',
+            None,
+            None,
+            'https://module-provider.com/{namespace}/{module}-{provider}/browse',
+            None,
+            'https://module-version.com/{namespace}/{module}-{provider}/browse',
+            False,
+            False,
+            None
+        ),
+        # Test with no URLs configured and all custom URLs enabled and not git provider
+        # - base URL
+        (
+            '/modules/moduledetails/fullypopulated/testprovider',
+            None,
+            None,
+            None,
+            None,
+            None,
+            True,
+            True,
+            None
+        ),
+        # - example
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/example/examples/test-example',
+            None,
+            None,
+            None,
+            None,
+            None,
+            True,
+            True,
+            None
+        ),
+        # - submodule
+        (
+            '/modules/moduledetails/fullypopulated/testprovider/1.5.0/submodule/modules/example-submodule1',
+            None,
+            None,
+            None,
+            None,
+            None,
+            True,
+            True,
+            None
+        ),
+    ])
+    def test_source_code_urls(self, url, git_provider_id,
+                              module_provider_browse_url_template, module_provider_base_url_template,
+                              module_version_browse_url_template, module_version_base_url_template,
+                              allow_custom_git_urls_module_provider, allow_custom_git_urls_module_version,
+                              expected_source):
+        """Test source code URL shown."""
+
+        module_provider = ModuleProvider.get(Module(Namespace('moduledetails'), 'fullypopulated'), 'testprovider')
+        module_version = ModuleVersion.get(module_provider, '1.5.0')
+        module_provider_row = module_provider._get_db_row()
+        module_version_row = module_version._get_db_row()
+
+        original_git_provider_id = module_provider_row['git_provider_id']
+        original_module_provider_browse_url_template = module_provider_row['repo_browse_url_template']
+        original_module_provider_base_url_template = module_provider_row['repo_base_url_template']
+        original_module_version_browse_url_template = module_version_row['repo_browse_url_template']
+        original_module_version_base_url_template = module_version_row['repo_base_url_template']
+
+
+        try:
+            module_version.update_attributes(
+                repo_browse_url_template=module_version_browse_url_template,
+                repo_base_url_template=module_version_base_url_template)
+            module_provider.update_attributes(
+                git_provider_id=git_provider_id,
+                repo_browse_url_template=module_provider_browse_url_template,
+                repo_base_url_template=module_provider_base_url_template
+            )
+
+            with self.update_mock(self._config_allow_custom_repo_urls_module_provider, 'new', allow_custom_git_urls_module_provider), \
+                    self.update_mock(self._config_allow_custom_repo_urls_module_version, 'new', allow_custom_git_urls_module_version):
+
+                self.selenium_instance.get(self.get_url(url))
+
+                # Ensure source code URL is correct
+                source_url = self.wait_for_element(By.ID, 'source-url', ensure_displayed=False)
+                if expected_source:
+                    self.assert_equals(lambda: source_url.is_displayed(), True)
+                    assert source_url.text == f'Source code: {expected_source}'
+                else:
+                    # Wait for inputs tab
+                    self.wait_for_element(By.ID, 'module-tab-link-inputs')
+                    assert source_url.is_displayed() == False
+
+        finally:
+            module_version.update_attributes(
+                repo_browse_url_template=original_module_version_browse_url_template,
+                repo_base_url_template=original_module_version_base_url_template)
+            module_provider.update_attributes(
+                git_provider_id=original_git_provider_id,
+                repo_browse_url_template=original_module_provider_browse_url_template,
+                repo_base_url_template=original_module_provider_base_url_template
+            )
 
     @pytest.mark.parametrize('url,expected_readme_content', [
         # Root module
@@ -973,7 +1434,7 @@ module "fullypopulated" {{
             # Check git provider dropdown
             git_provider_select_element = self.selenium_instance.find_element(By.ID, 'settings-git-provider')
 
-            expected_git_providers = ['testgitprovider', 'repo_url_tests', 'repo_url_tests_uri_encoded']
+            expected_git_providers = ['testgitprovider', 'repo_url_tests', 'repo_url_tests_uri_encoded', 'no_browse_url']
             if allow_custom_git_url_setting:
                 expected_git_providers.insert(0, 'Custom')