+++ test/selenium/test_module_search.py
@@ -35,6 +35,8 @@ class TestModuleSearch(SeleniumTest):
 
         self.assert_equals(lambda: self.selenium_instance.current_url, self.get_url('/modules/search?q=modulesearch'))
 
+        assert self.selenium_instance.title == 'Search - Terrareg'
+
         self.assert_equals(lambda: len(self.selenium_instance.find_element(By.ID, 'results').find_elements(By.CLASS_NAME, 'card')), 4)
 
     def test_result_cards(self):